RUB-250: avoid reclassifying version helper debt
